refactor!: Separate legacy MCP wiring from ActorsMcpServer
Extract all v1 MCP SDK wiring out of ActorsMcpServer into a new
package-private LegacyMcpServer (src/mcp/legacy_server.ts). ActorsMcpServer
becomes the shared-Apify facade that constructs and delegates to the adapter;
shared tool execution now takes plain values instead of the server object or
the v1 RequestHandlerExtra/McpError. This fixes the dependency direction PR
#1143 exposed, so a later stateless adapter can reuse one Apify core.

Behavior is unchanged: tool names, results, error codes/messages/data,
notifications, Tasks, telemetry, and timing are byte-for-byte identical,
verified by a live mcpc diff against the pre-change server plus the full unit
suite (1177 passed + 1 skipped). close() teardown is the deliberate
reverse-of-connect order (server close, then tool-map clear).

BREAKING: the raw v1 .server and .taskStore are no longer public on
ActorsMcpServer; they move to the private adapter. Runtime-safe for the hosted
server, which reads neither, but external TypeScript consumers touching them
break at compile time.

Closes #1147
Part of #1128
