+++ tests/unit/tool_call_engine.test.ts
@@ -1,22 +1,29 @@
-import type { RequestHandlerExtra } from '@modelcontextprotocol/sdk/shared/protocol.js';
-import type { Notification, Request } from '@modelcontextprotocol/sdk/types.js';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import log from '@apify/log';
 
 import { FAILURE_CATEGORY, TOOL_STATUS } from '../../src/const.js';
+import type { ActorsMcpServer } from '../../src/mcp/server.js';
 import type { InvalidToolCall, PreparedCall } from '../../src/mcp/tool_call_engine.js';
 import { executeSyncToolCall, prepareToolCall } from '../../src/mcp/tool_call_engine.js';
 import type { ToolCallTelemetryProperties } from '../../src/types.js';
 import { makePaymentRequiredError, makeRecorderTool, makeThrowingTool, withServer } from './helpers/mcp_server.js';
 
-/** Builds request context for direct engine tests. */
-function fakeExtra(aborted = false): RequestHandlerExtra<Request, Notification> {
+/** An abort signal for direct engine tests, optionally already aborted. */
+function makeSignal(aborted = false): AbortSignal {
+    const controller = new AbortController();
+    if (aborted) controller.abort();
+    return controller.signal;
+}
+
+/** Server-derived plain values that `prepareToolCall` reads. */
+function prepareFields(server: ActorsMcpServer) {
     return {
-        signal: { aborted },
-        sendNotification: vi.fn(),
-        requestInfo: { headers: {} },
-    } as unknown as RequestHandlerExtra<Request, Notification>;
+        tools: server.tools,
+        paymentProvider: server.options.paymentProvider,
+        allowUnauthMode: server.options.allowUnauthMode,
+        loadedToolNames: server.listToolNames(),
+    };
 }
 
 /** Suppresses expected failure logs. */
@@ -33,7 +40,7 @@ describe('prepareToolCall()', () => {
     it('returns InvalidToolCall with AUTH category when no token is provided', async () => {
         await withServer(async (server) => {
             const result = await prepareToolCall({
-                apifyMcpServer: server,
+                ...prepareFields(server),
                 apifyToken: '',
                 name: 'anything',
                 args: {},
@@ -55,7 +62,7 @@ describe('prepareToolCall()', () => {
     it('returns InvalidToolCall for an unknown tool name', async () => {
         await withServer(async (server) => {
             const result = await prepareToolCall({
-                apifyMcpServer: server,
+                ...prepareFields(server),
                 apifyToken: 'fake-token',
                 name: 'does-not-exist',
                 args: {},
@@ -78,7 +85,7 @@ describe('prepareToolCall()', () => {
             const { tool } = makeRecorderTool('recorder-tool');
             server.upsertTools([tool]);
             const result = await prepareToolCall({
-                apifyMcpServer: server,
+                ...prepareFields(server),
                 apifyToken: 'fake-token',
                 name: 'recorder-tool',
                 args: undefined,
@@ -101,7 +108,7 @@ describe('prepareToolCall()', () => {
             server.upsertTools([tool]);
             const telemetryData = { tool_name: 'recorder-tool' } as unknown as ToolCallTelemetryProperties;
             const result = await prepareToolCall({
-                apifyMcpServer: server,
+                ...prepareFields(server),
                 apifyToken: 'fake-token',
                 name: 'recorder-tool',
                 args: {},
@@ -125,10 +132,10 @@ describe('prepareToolCall()', () => {
 describe('executeSyncToolCall()', () => {
     afterEach(() => vi.restoreAllMocks());
 
-    async function prepare(server: Parameters<Parameters<typeof withServer>[0]>[0], name: string, tool: unknown) {
+    async function prepare(server: ActorsMcpServer, name: string, tool: unknown) {
         server.upsertTools([tool as never]);
         const prepared = (await prepareToolCall({
-            apifyMcpServer: server,
+            ...prepareFields(server),
             apifyToken: 'fake-token',
             name,
             args: {},
@@ -142,18 +149,28 @@ describe('executeSyncToolCall()', () => {
         return prepared;
     }
 
+    /** Server-derived plain values that `executeSyncToolCall` reads. */
+    function syncParams(server: ActorsMcpServer, toolName: string, aborted = false) {
+        return {
+            apifyToken: 'fake-token',
+            toolName,
+            mcpSessionId: 's1',
+            progressToken: undefined,
+            tools: server.tools,
+            actorStore: server.actorStore,
+            paymentProvider: server.options.paymentProvider,
+            loadedToolNames: server.listToolNames(),
+            signal: makeSignal(aborted),
+            sendNotification: vi.fn(),
+            emitLog: vi.fn(),
+        };
+    }
+
     it('returns a success outcome for a normal dispatch', async () => {
         await withServer(async (server) => {
             const { tool } = makeRecorderTool('recorder-tool');
             const prepared = await prepare(server, 'recorder-tool', tool);
-            const outcome = await executeSyncToolCall(prepared, {
-                apifyMcpServer: server,
-                apifyToken: 'fake-token',
-                toolName: 'recorder-tool',
-                mcpSessionId: 's1',
-                progressToken: undefined,
-                extra: fakeExtra(),
-            });
+            const outcome = await executeSyncToolCall(prepared, syncParams(server, 'recorder-tool'));
             expect(outcome.toolStatus).toBe(TOOL_STATUS.SUCCEEDED);
             expect((outcome.result.content as { text: string }[])[0].text).toBe('ok');
         });
@@ -164,14 +181,7 @@ describe('executeSyncToolCall()', () => {
             silenceLogs();
             const tool = makeThrowingTool({ name: 'payment-throwing-tool', error: makePaymentRequiredError() });
             const prepared = await prepare(server, 'payment-throwing-tool', tool);
-            const outcome = await executeSyncToolCall(prepared, {
-                apifyMcpServer: server,
-                apifyToken: 'fake-token',
-                toolName: 'payment-throwing-tool',
-                mcpSessionId: 's1',
-                progressToken: undefined,
-                extra: fakeExtra(),
-            });
+            const outcome = await executeSyncToolCall(prepared, syncParams(server, 'payment-throwing-tool'));
             expect(outcome.toolStatus).toBe(TOOL_STATUS.SOFT_FAIL);
             expect(outcome.callDiagnostics.failure_http_status).toBe(402);
             expect(outcome.result.isError).toBe(true);
@@ -183,14 +193,7 @@ describe('executeSyncToolCall()', () => {
             silenceLogs();
             const tool = makeThrowingTool({ name: 'aborting-tool', error: new Error('boom') });
             const prepared = await prepare(server, 'aborting-tool', tool);
-            const outcome = await executeSyncToolCall(prepared, {
-                apifyMcpServer: server,
-                apifyToken: 'fake-token',
-                toolName: 'aborting-tool',
-                mcpSessionId: 's1',
-                progressToken: undefined,
-                extra: fakeExtra(true),
-            });
+            const outcome = await executeSyncToolCall(prepared, syncParams(server, 'aborting-tool', true));
             expect(outcome.toolStatus).toBe(TOOL_STATUS.ABORTED);
             expect(outcome.result.isError).toBe(true);
             expect(outcome.result.toolTelemetry).toEqual({ toolStatus: TOOL_STATUS.ABORTED });
